more types cleanup
